python/examples/advertisements-agent-with-strands-agents-x402-cdp-chatbot/pyproject.toml
[project]
name = "ads-agent"
version = "0.0.1"
description = "CDP AgentKit x402 Example Chatbot"
authors = [{ name = "Uchenna Egbe", email = "ucheegbe@amazon.com" }]
requires-python = "~=3.10.0"
readme = "README.md"

dependencies = [
    "python-dotenv>=1.0.1,<2",
    "coinbase-agentkit",
    "coinbase-agentkit-strands-agents",
    "typing-extensions>=4.13.2",
    "fastapi",
    "uvicorn"
]

[dependency-groups]
dev = ["ruff>=0.7.1,<0.8"]

[tool.uv]
package = false
override-dependencies = [
    "typing-extensions>=4.13.2",
]
resolution = "highest"

[tool.uv.sources]
coinbase-agentkit = { path = "../../../python/coinbase-agentkit", editable = true }
coinbase-agentkit-strands-agents = { path = "../../../python/framework-extensions/strands-agents", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.advertisements-agent-with-strands-agents-x402-cdp-chatbot]
private = true

[tool.ruff]
line-length = 100
target-version = "py310"
exclude = ["./build/**", "./dist/**", "./docs/**"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "D", "UP", "B", "C4", "SIM", "RUF"]
ignore = ["D213", "D203", "D100", "D104", "D107", "E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["coinbase_agentkit, cdp"]